timeline component, dashboard refactor
timeline component displays state and triggers a callback to refresh the gui after one stage ends. it also has a transition stage which is when the refresh timer starts ( it checks if the stage is updated yet and refreshed the gui if so).
Dashboard now displays collab minimal info in the lists. Had to divide collab entity into metadata (light) and details (heavy, including submission list) so the info for multiple collabs can be fetched to the system efficiently. yeah, nice.